Change return type to Response of BaseTodoResponse

+++ app/src/main/java/com/example/innovateapprecruitmenttest/domain/repository/TodoRepositoryImpl.kt
@@ -5,6 +5,7 @@ import com.example.innovateapprecruitmenttest.di.API_KEY
 import com.example.innovateapprecruitmenttest.model.TodoListItem
 import com.example.innovateapprecruitmenttest.model.api.TodoAPI
 import com.example.innovateapprecruitmenttest.model.room.TodoDao
+import com.example.innovateapprecruitmenttest.utils.apiRequestError
 import com.example.innovateapprecruitmenttest.utils.handleResult
 import io.reactivex.disposables.CompositeDisposable
 
@@ -18,10 +19,10 @@ class TodoRepositoryImpl(
     val disposables = CompositeDisposable()
 
     override suspend fun getTodos(): List<TodoListItem> {
-        /**
-         * If there's no internet connection, default to the cached values.
-         * Otherwise propagate the error.
-         * */
+
+//         If there's no internet connection, default to the cached values.
+//         Otherwise propagate the error.
+
         val cachedTodos = todoDao.getSavedTodos()
         val apiTodos = try { // apiTodo returns RawTodo, so we must transform it to TodoListItem as follows:
             todoApi.getAllTodos(API_KEY).todos.map {
@@ -49,13 +50,19 @@ class TodoRepositoryImpl(
 
     override suspend fun insertTodo(todo: TodoListItem): TodoListItem? =
         try {
-            val apiIdResult = todoApi.insertTodo(API_KEY, todo.title).id
-            val updatedTodo = updateTodo(apiIdResult, todo)
-            Log.i("Updating todo result:", "Success:::${updatedTodo.toString()}")
-            if (updatedTodo != null) {
-                todoDao.insertTodo(updatedTodo)
-                updatedTodo
+            val apiRequest = todoApi.insertTodo(API_KEY, todo.title)
+            if (apiRequest.isSuccessful) {
+                val apiIdResult = apiRequest.body()!!.id
+                val updatedTodo = updateTodo(apiIdResult, todo)
+                Log.i("Updating todo result:", "Success:::${updatedTodo.toString()}")
+                if (updatedTodo != null) {
+                    todoDao.insertTodo(updatedTodo)
+                    updatedTodo
+                } else {
+                    null
+                }
             } else {
+                apiRequestError("inserting")
                 null
             }
         } catch (error: Throwable) {
@@ -65,10 +72,15 @@ class TodoRepositoryImpl(
 
 
     override suspend fun deleteTodo(todo: TodoListItem): Boolean = try {
-        todoApi.deleteTodo(API_KEY, todo.id)
-        todoDao.deleteTodo(todo)
-        Log.i("Deleting todo result:", "Success:::deleted ${todo.id}")
-        true
+        val apiRequest = todoApi.deleteTodo(API_KEY, todo.id)
+        if (apiRequest.isSuccessful) {
+            todoDao.deleteTodo(todo)
+            Log.i("Deleting todo result:", "Success:::deleted ${todo.id}")
+            true
+        } else {
+            apiRequestError("deleting")
+            false
+        }
     } catch (error: Throwable) {
         Log.i("Deleting todo result:", "Error:::${error.message}")
         false
@@ -81,14 +93,19 @@ class TodoRepositoryImpl(
             params["priority"] = todo.priority
             todo.description?.let { params["description"] = todo.description }
             todo.deadlineAt?.let { params["deadline_at"] = todo.deadlineAt }
-            todoApi.updateTodo(API_KEY, id, params)
-            TodoListItem(
-                id,
-                todo.title,
-                todo.description,
-                todo.priority,
-                todo.deadlineAt
-            )
+            val apiRequest = todoApi.updateTodo(API_KEY, id, params)
+            if (apiRequest.isSuccessful) {
+                TodoListItem(
+                    id,
+                    todo.title,
+                    todo.description,
+                    todo.priority,
+                    todo.deadlineAt
+                )
+            } else {
+                apiRequestError("updating")
+                null
+            }
         } catch (error: Throwable) {
             Log.i("Updating todo result:", "Error:::${error.message}")
             null
